requirements.txt
astroplan
astropy
astroquery
django
django-crispy-forms
django-extensions
django-filter
django-guardian
django-tasks
healpix-alchemy
healpy
lightcurve-fitting
ligo.skymap
matplotlib
numpy
tomtoolkit==2.24.5
tom-alertstreams
tom-nonlocalizedevents @ git+https://github.com/TOMToolkit/tom_nonlocalizedevents.git
plotly
psycopg2-binary
python-dateutil
requests
slack-sdk
specutils
SQLAlchemy
whitenoise
